refactor: extract useDeadlinePriority hook
